Renomme DownloadButton en DownloadCanvasButton
